+++ Assets/VRM/Editor/Format/VRMMaterialValidator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 43a9939275e36534797e9be6fb83c71e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
