+++ Assets/VRM10/Editor/VRM10MaterialValidator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77bcf6c00ae8990488aaea60f0ec9173
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
